+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show tiles cache size in region management UI
 - Online Nominatim geocoding
 - Zoom-in using double-tap
+- Map marker for search results
 
 ### Changed